requirements.txt
torch==1.12.0+cu116
transformers==4.25.0
scikit-learn==0.22
pandas==1.3.5
torchvision==0.16.0
Pillow==10.0.1


